Split Hugging Face realtime tests
